refactor(toolpaths): extract shared trochoidal entry maths into trochoidalPath.ts
Move TROCHOIDAL_ENTRY_STEPS_PER_REVOLUTION, MAX_TROCHOIDAL_ENTRY_MOVES,
TrochoidalOperationBudget, trochoidalEntryStrategy, appendTrochoidalEntry,
trochoidalEntryPoints, and trochoidalEntryMoveCount from edge.ts into a new
trochoidalPath.ts module. Zero behaviour change — bodies are verbatim with
export added. Edge Route imports them back; Engrave will consume the same
clearance maths in a later slice.
