+++ src/engine/toolpaths/INDEX.md
@@ -26,6 +26,7 @@ Toolpath generators. Each file owns one strategy. `index.ts` re-exports everythi
 - `feed.ts` — shared `effectiveFeed` helper: applies a move's `feedScale` (for slot-feed pocket cuts) to the cut feed, and returns the plunge feed unmodified for plunge moves; used by the postprocessor, booklet time estimator, simulation playback, and the live-feed readout
 - `entry.ts` — clearance-aware plunge, helical, and zig-zag ramp entry synthesis for pocket, surface-clean, and rough-surface clearing operations, including deterministic fallback warnings and plunge-feed limiting
 - `geometry.ts` — toolpath-specific geometric helpers; owns the shared `DEFAULT_FLATTEN_*` sampling constants
+- `trochoidalPath.ts` — shared trochoidal entry synthesis (helix/plunge points, move-count budgeting, and the per-operation point budget type) used by both Edge Route and, later, Engrave
 - `guideFragments.ts` — cyclic closed-guide fragmentation against pre-unioned keep-outs; preserves every safe span before entry-bearing motion is generated
 - `tabs.ts` also owns the shared tab footprint geometry (`expandedTabFootprints`) and `applyEdgeRouteTabs`, the edge-route tab pass that deliberately returns trochoidal results untouched — see `planning/TROCHOIDAL_EDGE_DESIGN.md`
 - `offsetSmoothing.ts` — emit-time corner fillet for the outer/wall clearing rings (`roundContourCorners`, `smoothClosedContours`, `cornerSmoothingRadius`); shared by pocket + surface clearing when `roundOutsideCorners` is enabled. Bounds the setback so acute corners leave no crescent; the offset-tree emitter keeps each region's wall-adjacent (root) outer ring sharp so no corner stock stacks into a chip (interior rings self-clean). Islands are rounded the opposite way — via `jtRound` Clipper offsets (see `buildInsetRegions` island join), so the tool wraps convex island corners smoothly without gouging.
