+++ src/engine/toolpaths/index.ts
@@ -36,6 +36,7 @@ export * from './surfaceStepdown3d'
 export * from './surface'
 export * from './tabs'
 export * from './trochoidalEdge'
+export * from './trochoidalPath'
 export * from './types'
 export * from './vcarve'
 export * from './vcarveMedial'
